+++ src/arcemu-world/ArathiBasin.cpp
@@ -226,7 +226,7 @@ void ArathiBasin::SpawnControlPoint(uint32 Id, uint32 Type)
 		m_controlPoints[Id]->SetFloatValue(GAMEOBJECT_PARENTROTATION_02, ControlPointRotations[Id][0]);
 		m_controlPoints[Id]->SetFloatValue(GAMEOBJECT_PARENTROTATION_03, ControlPointRotations[Id][1]);
 		m_controlPoints[Id]->SetByte(GAMEOBJECT_BYTES_1, 0, 1);
-		m_controlPoints[Id]->SetByte(GAMEOBJECT_BYTES_1, 1, gi->Type);
+		m_controlPoints[Id]->SetByte(GAMEOBJECT_BYTES_1, 1, static_cast<uint8>( gi->Type ));
 		m_controlPoints[Id]->SetByte(GAMEOBJECT_BYTES_1, 3, 100);
 		m_controlPoints[Id]->SetUInt32Value(GAMEOBJECT_DYNAMIC, 1);
 		m_controlPoints[Id]->SetUInt32Value(GAMEOBJECT_DISPLAYID, gi->DisplayID);
@@ -248,7 +248,7 @@ void ArathiBasin::SpawnControlPoint(uint32 Id, uint32 Type)
 			break;
 		}
 
-		m_controlPoints[Id]->bannerslot = Id;
+		m_controlPoints[Id]->bannerslot = static_cast<uint8>( Id );
 		m_controlPoints[Id]->PushToWorld(m_mapMgr);
 	}
 	else
@@ -260,7 +260,7 @@ void ArathiBasin::SpawnControlPoint(uint32 Id, uint32 Type)
 		m_controlPoints[Id]->SetNewGuid(m_mapMgr->GenerateGameobjectGuid());
 		m_controlPoints[Id]->SetUInt32Value(OBJECT_FIELD_ENTRY, gi->ID);
 		m_controlPoints[Id]->SetUInt32Value(GAMEOBJECT_DISPLAYID, gi->DisplayID);
-		m_controlPoints[Id]->SetByte(GAMEOBJECT_BYTES_1, 1, gi->Type);
+		m_controlPoints[Id]->SetByte(GAMEOBJECT_BYTES_1, 1, static_cast<uint8>( gi->Type ));
 
 		switch(Type)
 		{
@@ -302,7 +302,7 @@ void ArathiBasin::SpawnControlPoint(uint32 Id, uint32 Type)
 		m_controlPointAuras[Id]->SetByte(GAMEOBJECT_BYTES_1, 0, 1);
 		m_controlPointAuras[Id]->SetByte(GAMEOBJECT_BYTES_1, 1, 6);
 		m_controlPointAuras[Id]->SetByte(GAMEOBJECT_BYTES_1, 3, 100);
-		m_controlPointAuras[Id]->bannerauraslot = Id;
+		m_controlPointAuras[Id]->bannerauraslot = static_cast<uint8>( Id );
 		m_controlPointAuras[Id]->PushToWorld(m_mapMgr);
 	}
 	else
@@ -588,7 +588,7 @@ void ArathiBasin::EventUpdateResources(uint32 Team)
 	if(current_resources == RESOURCES_WINVAL)
 	{
 		m_ended = true;
-		m_winningteam = Team;
+		m_winningteam = static_cast<uint8>( Team );
 		m_nextPvPUpdateTime = 0;
 
 		sEventMgr.RemoveEvents(this);
