+++ src/arcemu-world/Item.cpp
@@ -428,7 +428,7 @@ void Item::DeleteFromDB()
 		/* deleting a container */
 		for( uint32 i = 0; i < m_itemProto->ContainerSlots; ++i )
 		{
-			if( static_cast< Container* >( this )->GetItem( i ) != NULL )
+			if( static_cast< Container* >( this )->GetItem( static_cast<int16>( i ) ) != NULL )
 			{
 				/* abort the delete */
 				return;
