+++ src/arcemu-world/Level2.cpp
@@ -1149,11 +1149,11 @@ bool ChatHandler::HandleAddAIAgentCommand(const char* args, WorldSession *m_sess
 	WorldDatabase.Execute( qry.str().c_str( ) );
 
 	AI_Spell * sp = new AI_Spell;
-	sp->agent = atoi(agent);
+	sp->agent = static_cast<uint16>( atoi(agent) );
 	sp->procChance = atoi(procChance);
 /*	sp->procCount = atoi(procCount);*/
 	sp->spell = dbcSpell.LookupEntry(atoi(spellId));
-	sp->spellType = atoi(spellType);
+	sp->spellType = static_cast<uint8>( atoi(spellType) );
 //	sp->spelltargetType = atoi(spelltargetType);
 	sp->floatMisc1 = (float)atof(floatMisc1);
 	sp->Misc2 = (uint32)atof(Misc2);
@@ -1222,7 +1222,7 @@ bool ChatHandler::HandleGOAnimProgress(const char * args, WorldSession * m_sessi
 		return false;
 
 	uint32 ap = atol(args);
-	GObj->SetByte( GAMEOBJECT_BYTES_1, 3, ap);
+	GObj->SetByte( GAMEOBJECT_BYTES_1, 3, static_cast<uint8>( ap ));
 	BlueSystemMessage(m_session, "Set ANIMPROGRESS to %u", ap);
 	return true;
 }
@@ -1395,7 +1395,7 @@ bool ChatHandler::HandleRepairItemsCommand(const char *args, WorldSession *m_ses
 
 	for( i = 0; i < MAX_INVENTORY_SLOT; i++ )
 	{
-		pItem = plr->GetItemInterface()->GetInventoryItem( i );
+		pItem = plr->GetItemInterface()->GetInventoryItem( static_cast<uint16>( i ) );
 		if( pItem != NULL )
 		{
 			if( pItem->IsContainer() )
@@ -1403,7 +1403,7 @@ bool ChatHandler::HandleRepairItemsCommand(const char *args, WorldSession *m_ses
 				pContainer = static_cast<Container*>( pItem );
 				for( j = 0; j < pContainer->GetProto()->ContainerSlots; ++j )
 				{
-					pItem = pContainer->GetItem( j );
+					pItem = pContainer->GetItem( static_cast<uint16>( j ) );
 					if( pItem != NULL )
 						RepairItem2( plr, pItem );
 				}
@@ -1413,7 +1413,7 @@ bool ChatHandler::HandleRepairItemsCommand(const char *args, WorldSession *m_ses
 				if( pItem->GetProto()->MaxDurability > 0 && i < INVENTORY_SLOT_BAG_END && pItem->GetDurability() <= 0 )
 				{
 					RepairItem2( plr, pItem );
-					plr->ApplyItemMods( pItem, i, true );
+					plr->ApplyItemMods( pItem, static_cast<uint16>( i ), true );
 				}
 				else
 				{
